% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg1d.R
\name{beautify}
\alias{beautify}
\title{Make strings prettier for printing}
\usage{
beautify(string, autodetect_units = TRUE)
}
\arguments{
\item{string}{input string}

\item{autodetect_units}{automatically detect units (e.g. mm, kg, etc) and wrap in brackets.}
}
\value{
string
}
\description{
Takes an input string and 'beautify' by converting underscores to spaces and
}
