% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_df.R
\name{column_info_table}
\alias{column_info_table}
\title{Parse a tibble and ensure it meets standards}
\usage{
column_info_table(.data, maxlevels = 6, col_id)
}
\arguments{
\item{.data}{data.frame to autoplot (data.frame)}

\item{maxlevels}{for categorical variables, what is the maximum number of distinct values to allow (too many will make it hard to find a palette that suits). (number)}

\item{col_id}{name of column to use for}
}
\value{
tibble with the following columns:
\enumerate{
\item colnames
\item coltype (categorical/numeric/tooltip/invalid)
\item ndistinct (number of distinct values)
\item plottable (should this column be plotted)
}
}
\description{
Parse a tibble and ensure it meets standards
}
\examples{
path_gg1d <- system.file('testdata/testinput.csv', package = 'gg1d')
df <- read.csv(path_gg1d, header = TRUE)
gg1d:::column_info_table(df)
}
