% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparallel_specific_utils.R
\name{fct_infreq}
\alias{fct_infreq}
\title{Reorder Factor Levels by Descending Frequency}
\usage{
fct_infreq(x)
}
\arguments{
\item{x}{A factor or an object coerced to a factor.}
}
\value{
A factor with levels ordered by descending frequency.
}
\description{
Reorders the levels of a factor by their frequency, in descending order.
}
\keyword{internal}
