% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparallel_specific_utils.R
\name{fct_relevel_base}
\alias{fct_relevel_base}
\title{Relevel Factor by Specified Levels}
\usage{
fct_relevel_base(x, ..., after = 0)
}
\arguments{
\item{x}{A factor to be releveled.}

\item{...}{Levels to move in the factor.}

\item{after}{A numeric scalar specifying the position after which the moved
levels should be placed. Use \code{0} to place them at the front.}
}
\value{
A factor with the specified levels moved to the chosen position.
}
\description{
Reorder the levels of a factor by moving specified levels to a new position.
}
\keyword{internal}
