% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparallel_specific_utils.R
\name{fct_rev}
\alias{fct_rev}
\title{Reverse the Levels of a Factor}
\usage{
fct_rev(x)
}
\arguments{
\item{x}{A factor or an object coerced to a factor.}
}
\value{
A factor with reversed levels.
}
\description{
Reverses the existing level order of a factor.
}
\keyword{internal}
