% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparallel_specific_utils.R
\name{feature_vector_to_total_path_distance}
\alias{feature_vector_to_total_path_distance}
\title{Compute the Total Path Distance for an Axis Order}
\usage{
feature_vector_to_total_path_distance(axis_names, mx)
}
\arguments{
\item{axis_names}{A character vector indicating the axis order.}

\item{mx}{A matrix of distances, with row and column names matching \code{axis_names}.}
}
\value{
A numeric value representing the total distance.
}
\description{
Given a sequence of axis names and a distance matrix, sums pairwise distances
along the path.
}
\keyword{internal}
