% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg1d-package.R
\docType{package}
\name{gg1d-package}
\alias{gg1d-package}
\title{gg1d: Exploratory Data Analysis using Tiled One-Dimensional Graphics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Streamlines exploratory data analysis by providing a turnkey approach to visualising n-dimensional data which graphically reveals correlative or associative relationships between 2 or more features. Represents all dataset features as distinct, vertically aligned bar or tile plots, with plot types auto-selected based on whether variables are categorical or numeric.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/selkamand/gg1d}
  \item \url{https://selkamand.github.io/gg1d/}
  \item Report bugs at \url{https://github.com/selkamand/gg1d/issues}
}

}
\author{
\strong{Maintainer}: Sam El-Kamand \email{sam.elkamand@gmail.com} (\href{https://orcid.org/0000-0003-2270-8088}{ORCID})

Other contributors:
\itemize{
  \item Children's Cancer Institute Australia [copyright holder]
}

}
\keyword{internal}
