% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_df.R
\name{gg1d_plot}
\alias{gg1d_plot}
\title{AutoPlot an entire data.frame}
\usage{
gg1d_plot(
  .data,
  col_id = NULL,
  col_sort_by = character(0),
  maxlevels = 6,
  verbose = TRUE,
  debug_return_col_info = FALSE
)
}
\arguments{
\item{.data}{data.frame to autoplot (data.frame)}

\item{col_id}{name of column to use for}

\item{col_sort_by}{name of column to sort on}

\item{maxlevels}{for categorical variables, what is the maximum number of distinct values to allow (too many will make it hard to find a palette that suits). (number)}

\item{verbose}{verbosity level}
}
\value{
ggiraph interactive visualisation
}
\description{
Visualize all columns in a data frame with gg1d's vertically aligned plots
and automatic plot selection based on variable type.
Plots are fully interactive, and custom tooltips can be added.
}
\examples{
path_gg1d <- system.file('testdata/testinput.csv', package = 'gg1d')
df <- read.csv(path_gg1d, header = TRUE)
gg1d_plot(df)


}
