% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparallel_specific_utils.R
\name{is_crossing}
\alias{is_crossing}
\title{Determine Whether Two Edges Cross}
\usage{
is_crossing(l1, r1, l2, r2)
}
\arguments{
\item{l1}{Numeric position of the first edge on the left axis.}

\item{r1}{Numeric position of the first edge on the right axis.}

\item{l2}{Numeric position of the second edge on the left axis.}

\item{r2}{Numeric position of the second edge on the right axis.}
}
\value{
A logical value. \code{TRUE} if they cross, \code{FALSE} otherwise.
}
\description{
Given the positions of two edges on the left and right axes, decides if they
intersect in a parallel coordinates setup.
}
\keyword{internal}
