% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lazy_birdwatcher}
\alias{lazy_birdwatcher}
\title{Lazy Birdwatcher Dataset}
\format{
\subsection{\code{lazy_birdwatcher}}{

A data frame with 45 rows and 3 columns:

\describe{
\item{Magpies}{Number of magpies observed}
\item{Day}{Was the day of observation a weekday or a weekend?}
\item{Birdwatcher}{Name of the birdwatcher}
}
}
}
\usage{
lazy_birdwatcher
}
\description{
A simulated dataset describing the number of magpies observed by two birdwatchers.
}
\keyword{datasets}
