% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparallel_specific_utils.R
\name{permutations}
\alias{permutations}
\title{Generate Permutations of the Integers 1..n}
\usage{
permutations(n)
}
\arguments{
\item{n}{Number of elements to permute.}
}
\value{
A matrix where each row is a permutation of 1..n.
}
\description{
Creates a matrix of all permutations for the integers from 1 to n.
}
\keyword{internal}
