% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparallel_specific_utils.R
\name{permute_axis_names}
\alias{permute_axis_names}
\title{Generate All Permutations of Axis Names}
\usage{
permute_axis_names(axis_names)
}
\arguments{
\item{axis_names}{A character vector of axis names.}
}
\value{
A matrix where each row represents one permutation of \code{axis_names}.
}
\description{
Takes a character vector of axis names and returns a matrix of permutations.
}
\keyword{internal}
