% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg1d.R
\name{sensible_2_breaks}
\alias{sensible_2_breaks}
\title{GGplot breaks}
\usage{
sensible_2_breaks(vector)
}
\arguments{
\item{vector}{vector fed into ggplot axis you want to define sensible breaks for}
}
\value{
vector of length 2. first element descripts upper break position, lower describes lower break
}
\description{
Find sensible values to add 2 breaks at for a ggplot2 axis
}
