% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_df.R
\name{column_info_table}
\alias{column_info_table}
\title{Parse a tibble and ensure it meets standards}
\usage{
column_info_table(.data, maxlevels = 6, col_id = NULL, cols_to_plot)
}
\arguments{
\item{.data}{data.frame to autoplot (data.frame)}

\item{maxlevels}{for categorical variables, what is the maximum number of distinct values to allow (too many will make it hard to find a palette that suits). (number)}

\item{col_id}{name of column to use for}

\item{cols_to_plot}{names of columns in \code{.data} that should be plotted. By default plots all valid columns (character)}
}
\value{
tibble with the following columns:
\enumerate{
\item colnames
\item coltype (categorical/numeric/tooltip/invalid)
\item ndistinct (number of distinct values)
\item plottable (should this column be plotted)
}
}
\description{
Parse a tibble and ensure it meets standards
}
