% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_df.R
\name{column_info_table}
\alias{column_info_table}
\title{Parse a tibble and ensure it meets standards}
\usage{
column_info_table(
  .data,
  maxlevels = 6,
  col_id = NULL,
  cols_to_plot,
  tooltip_column_suffix = "_tooltip",
  palettes,
  colours_default,
  colours_default_logical
)
}
\arguments{
\item{.data}{data.frame to autoplot (data.frame)}

\item{maxlevels}{for categorical variables, what is the maximum number of distinct values to allow (too many will make it hard to find a palette that suits). (number)}

\item{col_id}{name of column to use for}

\item{cols_to_plot}{names of columns in \code{.data} that should be plotted. By default plots all valid columns (character)}

\item{tooltip_column_suffix}{the suffix added to a column name that indicates column should be used as a tooltip (string)}

\item{palettes}{A list of named vectors. List names correspond to .data column names (categorical only). Vector names to levels of columns. Vector values are colours, the vector names are used to map values in .data to a colour.}

\item{colours_default}{default colours to use for variables. will be used to colour variables with no palette supplied.}

\item{colours_default_logical}{colours for binary variables (vector of 3 colors where elements represent colours of TRUE, FALSE, and NA respectively) (character)}
}
\value{
tibble with the following columns:
\enumerate{
\item colnames
\item coltype (categorical/numeric/tooltip/invalid)
\item ndistinct (number of distinct values)
\item plottable (should this column be plotted)
\item tooltip_col (the name of the column to use as the tooltip) or NA if no obvious tooltip column found
}
}
\description{
Parse a tibble and ensure it meets standards
}
