% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg1d.R
\name{gg1d_plot}
\alias{gg1d_plot}
\title{AutoPlot an entire data.frame}
\usage{
gg1d_plot(
  .data,
  col_id = NULL,
  col_sort = NULL,
  maxlevels = 6,
  verbose = TRUE,
  drop_unused_id_levels = FALSE,
  interactive = TRUE,
  debug_return_col_info = FALSE,
  palettes = NULL,
  colours_default = c("#66C2A5", "#FC8D62", "#8DA0CB", "#E78AC3", "#A6D854", "#FFD92F",
    "#E5C494"),
  colours_default_logical = c(`TRUE` = "#648fff", `FALSE` = "#dc267f"),
  colours_missing = "grey90",
  limit_plots = TRUE,
  cols_to_plot = NULL,
  sort_type = c("frequency", "alphabetical"),
  desc = TRUE,
  width = 0.9,
  relative_height_numeric = 4,
  tooltip_column_suffix = "_tooltip",
  show_legend_titles = FALSE,
  show_legend = !interactive,
  legend_position = c("right", "left", "bottom", "top"),
  legend_title_position = c("top", "bottom", "left", "right"),
  legend_title_beautify = TRUE,
  legend_nrow = 4,
  legend_ncol = NULL,
  legend_title_size = NULL,
  legend_text_size = NULL,
  legend_key_size = 0.3
)
}
\arguments{
\item{.data}{data.frame to autoplot (data.frame)}

\item{col_id}{name of column to use for}

\item{col_sort}{column to sort sample order by. By default uses the supplied order of levels in col_id (order of appearance if a character type)}

\item{maxlevels}{for categorical variables, what is the maximum number of distinct values to allow (too many will make it hard to find a palette that suits). (number)}

\item{verbose}{verbosity level}

\item{drop_unused_id_levels}{if col_id is a factor with unused levels, should these be dropped or included in visualisation}

\item{interactive}{produce interactive ggiraph visualiastion (flag)}

\item{debug_return_col_info}{return column info instead of plots. Helpful when debugging (logical)}

\item{palettes}{A list of named vectors. List names correspond to .data column names (categorical only). Vector names to levels of columns. Vector values are colours, the vector names are used to map values in .data to a colour.}

\item{colours_default}{default colours to use for variables. will be used to colour variables with no palette supplied.}

\item{colours_default_logical}{colours for binary variables (vector of 3 colors where elements represent colours of TRUE, FALSE, and NA respectively) (character)}

\item{colours_missing}{colour to use for values of NA (string)}

\item{limit_plots}{throw an error when there are > 15 plottable columns in table (logical)}

\item{cols_to_plot}{names of columns in \code{.data} that should be plotted. By default plots all valid columns (character)}

\item{sort_type}{controls how categorical variables are sorted.
Numerical variables are always sorted in numerical order irrespective of the value given here.
Options are \code{alphabetical} or \code{frequency}}

\item{desc}{sort in descending order (flag)}

\item{width}{controls how much space is present between bars and tiles within each plot. Can be 0-1 where values of 1 makes bars/tiles take up 100\% of available space (no gaps between bars)}

\item{relative_height_numeric}{how many times taller should numeric plots be relative to categorical tile plots (number)}

\item{tooltip_column_suffix}{the suffix added to a column name that indicates column should be used as a tooltip (string)}

\item{show_legend_titles}{show legend titles (flag)}

\item{show_legend}{show the legend (flag)}

\item{legend_position}{position of the legend on the plot (string, options are "right", "left", "bottom", "top")}

\item{legend_title_position}{position of the title of the legend on the plot (string, options are "top", "bottom", "left", "right")}

\item{legend_title_beautify}{beautify legend title (add spaces to snake_case / camelCase & capitalise each word) (flag)}

\item{legend_nrow}{the number of rows in the legend (number)}

\item{legend_ncol}{the number of columns in the legend. Set \code{legend_nrow = NULL} when using legend_ncol (number)}

\item{legend_title_size}{the size of the title of the legend (number)}

\item{legend_text_size}{the size of the text in the legend (number)}

\item{legend_key_size}{the size of the key in the legend (number)}
}
\value{
ggiraph interactive visualisation
}
\description{
Visualize all columns in a data frame with gg1d's vertically aligned plots
and automatic plot selection based on variable type.
Plots are fully interactive, and custom tooltips can be added.
}
\examples{
path_gg1d <- system.file("testdata/testinput.csv", package = "gg1d")
df <- read.csv(path_gg1d, header = TRUE, na.strings = "")
gg1d_plot(df, col_id = "ID", col_sort = "Glasses")

}
